requirements.txt
fastapi
uvicorn
mongoengine
pymupdf
sentence-transformers
torch
openai
redis
python-dotenv